+++ example/lib/main.dart
@@ -15,10 +15,11 @@ class MyAppState extends State<MyApp> with WidgetsBindingObserver {
   late double windowHeight;
   late ZoomifyState zoomifyState;
   double currentZoomLevel = 1;
+  ZoomifyController zoomifyController = ZoomifyController();
 
 //  static const String folderUrl = 'https://kaartdekaag1933.zeilvaartwarmond.nl/kaartderkagerplassen-1933';
-  static const String folderUrl = 'https://chaerte.zeilvaartwarmond.nl/Warmond_J_Douw_1667';
 //  static const photoTitle = 'Kaart der Kagerplassen, Uitgave 1933';
+  static const String folderUrl = 'https://chaerte.zeilvaartwarmond.nl/Warmond_J_Douw_1667';
   static const photoTitle = 'Chaerte vande vrye Heerlickheydt Warmondt, Johannes Douw, 1667';
 
   GlobalKey zoomifyKey = GlobalKey();
@@ -50,7 +51,9 @@ class MyAppState extends State<MyApp> with WidgetsBindingObserver {
                 onChange: (zoomLevel, offset) => handleChange(zoomLevel, offset),
                 onImageReady: (width, height, maxZoom) => handleImageReady(width, height, maxZoom),
                 animationDuration: Duration(milliseconds: 500),
-                animationCurve: Curves.easeOut)));
+                animationCurve: Curves.easeOut,
+                animationSync: false,
+                controller: zoomifyController)));
   }
 
   void handleImageReady(int width, int height, int maxZoomLevel) {
@@ -63,35 +66,33 @@ class MyAppState extends State<MyApp> with WidgetsBindingObserver {
   }
 
   //
-  // Change zoom and pan programmatically. Instead of animateZoomAndPan, you can also use zoomAndPan
-  // you can also combine pan info and zoom info in on call
+  // Change zoom and pan programmatically. Instead of animateZoomAndPan, you can also use zoomAndPan for non-animated
+  // zooming and panning. Set the optional parameter sync to true to trigger the onChange callback function each animation frame.
+  // In principle you can combine pan info and zoom info in one call
+  //
+
   void zoomInOut(zoomLevelDelta) {
-    dynamic state = zoomifyKey.currentState;
-    state.animateZoomAndPan(zoomLevel: currentZoomLevel + zoomLevelDelta, zoomCenter: Offset(windowWidth / 2, windowHeight / 2));
+    zoomifyController.animateZoomAndPan(
+        zoomLevel: currentZoomLevel + zoomLevelDelta, zoomCenter: Offset(windowWidth / 2, windowHeight / 2));
   }
 
   void panUp() {
-    dynamic state = zoomifyKey.currentState;
-    state.animateZoomAndPan(panOffset: Offset(0, -100));
+    zoomifyController.animateZoomAndPan(panOffset: Offset(0, -100));
   }
 
   void panDown() {
-    dynamic state = zoomifyKey.currentState;
-    state.animateZoomAndPan(panOffset: Offset(0, 100));
+    zoomifyController.animateZoomAndPan(panOffset: Offset(0, 100));
   }
 
   void panLeft() {
-    dynamic state = zoomifyKey.currentState;
-    state.animateZoomAndPan(panOffset: Offset(-100, 0));
+    zoomifyController.animateZoomAndPan(panOffset: Offset(-100, 0));
   }
 
   void panRight() {
-    dynamic state = zoomifyKey.currentState;
-    state.animateZoomAndPan(panOffset: Offset(100, 0));
+    zoomifyController.animateZoomAndPan(panOffset: Offset(100, 0));
   }
 
   void reset() {
-    dynamic state = zoomifyKey.currentState;
-    state.reset();
+    zoomifyController.reset();
   }
 }